requirements.txt

fastapi
uvicorn
openai
requests
redis-om
rich
httpx
python-dotenv
asyncify
pillow
rich-traceback